+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Refactored `FileVersionInfo` to `FileVersion`
 * `ScanPoliciesManager` exclusion interface changed
 
+### Fixed
+* Fix call to incorrect internal api in `B2Api.get_download_url_for_file_name`
+
 ## [1.8.0] - 2021-05-21
 
 ### Added